+++ integrationtests/self/deadline_test.go
@@ -1,6 +1,7 @@
 package self_test
 
 import (
+	"context"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -27,9 +28,9 @@ var _ = Describe("Stream deadline tests", func() {
 		acceptedStream := make(chan struct{})
 		go func() {
 			defer GinkgoRecover()
-			sess, err := server.Accept()
+			sess, err := server.Accept(context.Background())
 			Expect(err).ToNot(HaveOccurred())
-			serverStr, err = sess.AcceptStream()
+			serverStr, err = sess.AcceptStream(context.Background())
 			Expect(err).ToNot(HaveOccurred())
 			_, err = serverStr.Read([]byte{0})
 			Expect(err).ToNot(HaveOccurred())