+++ integrationtests/self/handshake_drop_test.go
@@ -1,6 +1,7 @@
 package self_test
 
 import (
+	"context"
 	"fmt"
 	mrand "math/rand"
 	"net"
@@ -57,10 +58,10 @@ var _ = Describe("Handshake drop tests", func() {
 			serverSessionChan := make(chan quic.Session)
 			go func() {
 				defer GinkgoRecover()
-				sess, err := ln.Accept()
+				sess, err := ln.Accept(context.Background())
 				Expect(err).ToNot(HaveOccurred())
 				defer sess.Close()
-				str, err := sess.AcceptStream()
+				str, err := sess.AcceptStream(context.Background())
 				Expect(err).ToNot(HaveOccurred())
 				b := make([]byte, 6)
 				_, err = gbytes.TimeoutReader(str, 10*time.Second).Read(b)
@@ -92,7 +93,7 @@ var _ = Describe("Handshake drop tests", func() {
 			serverSessionChan := make(chan quic.Session)
 			go func() {
 				defer GinkgoRecover()
-				sess, err := ln.Accept()
+				sess, err := ln.Accept(context.Background())
 				Expect(err).ToNot(HaveOccurred())
 				str, err := sess.OpenStream()
 				Expect(err).ToNot(HaveOccurred())
@@ -106,7 +107,7 @@ var _ = Describe("Handshake drop tests", func() {
 				&quic.Config{Versions: []protocol.VersionNumber{version}},
 			)
 			Expect(err).ToNot(HaveOccurred())
-			str, err := sess.AcceptStream()
+			str, err := sess.AcceptStream(context.Background())
 			Expect(err).ToNot(HaveOccurred())
 			b := make([]byte, 6)
 			_, err = gbytes.TimeoutReader(str, 10*time.Second).Read(b)
@@ -126,7 +127,7 @@ var _ = Describe("Handshake drop tests", func() {
 			serverSessionChan := make(chan quic.Session)
 			go func() {
 				defer GinkgoRecover()
-				sess, err := ln.Accept()
+				sess, err := ln.Accept(context.Background())
 				Expect(err).ToNot(HaveOccurred())
 				serverSessionChan <- sess
 			}()