+++ streams_map_incoming_bidi.go
@@ -5,6 +5,7 @@
 package quic
 
 import (
+	"context"
 	"sync"
 
 	"github.com/lucas-clemente/quic-go/internal/protocol"
@@ -12,8 +13,8 @@ import (
 )
 
 type incomingBidiStreamsMap struct {
-	mutex sync.RWMutex
-	cond  sync.Cond
+	mutex         sync.RWMutex
+	newStreamChan chan struct{}
 
 	streams map[protocol.StreamNum]streamI
 	// When a stream is deleted before it was accepted, we can't delete it immediately.
@@ -36,9 +37,9 @@ func newIncomingBidiStreamsMap(
 	newStream func(protocol.StreamNum) streamI,
 	maxStreams uint64,
 	queueControlFrame func(wire.Frame),
-	// streamNumToID func(protocol.StreamNum) protocol.StreamID,
 ) *incomingBidiStreamsMap {
-	m := &incomingBidiStreamsMap{
+	return &incomingBidiStreamsMap{
+		newStreamChan:      make(chan struct{}),
 		streams:            make(map[protocol.StreamNum]streamI),
 		streamsToDelete:    make(map[protocol.StreamNum]struct{}),
 		maxStream:          protocol.StreamNum(maxStreams),
@@ -47,38 +48,43 @@ func newIncomingBidiStreamsMap(
 		nextStreamToOpen:   1,
 		nextStreamToAccept: 1,
 		queueMaxStreamID:   func(f *wire.MaxStreamsFrame) { queueControlFrame(f) },
-		// streamNumToID:      streamNumToID,
 	}
-	m.cond.L = &m.mutex
-	return m
 }
 
-func (m *incomingBidiStreamsMap) AcceptStream() (streamI, error) {
+func (m *incomingBidiStreamsMap) AcceptStream(ctx context.Context) (streamI, error) {
 	m.mutex.Lock()
-	defer m.mutex.Unlock()
 
 	var num protocol.StreamNum
 	var str streamI
 	for {
 		num = m.nextStreamToAccept
-		var ok bool
 		if m.closeErr != nil {
+			m.mutex.Unlock()
 			return nil, m.closeErr
 		}
+		var ok bool
 		str, ok = m.streams[num]
 		if ok {
 			break
 		}
-		m.cond.Wait()
+		m.mutex.Unlock()
+		select {
+		case <-ctx.Done():
+			return nil, ctx.Err()
+		case <-m.newStreamChan:
+		}
+		m.mutex.Lock()
 	}
 	m.nextStreamToAccept++
 	// If this stream was completed before being accepted, we can delete it now.
 	if _, ok := m.streamsToDelete[num]; ok {
 		delete(m.streamsToDelete, num)
 		if err := m.deleteStream(num); err != nil {
+			m.mutex.Unlock()
 			return nil, err
 		}
 	}
+	m.mutex.Unlock()
 	return str, nil
 }
 
@@ -111,7 +117,10 @@ func (m *incomingBidiStreamsMap) GetOrOpenStream(num protocol.StreamNum) (stream
 	// * highestStream is only modified by this function
 	for newNum := m.nextStreamToOpen; newNum <= num; newNum++ {
 		m.streams[newNum] = m.newStream(newNum)
-		m.cond.Signal()
+		select {
+		case m.newStreamChan <- struct{}{}:
+		default:
+		}
 	}
 	m.nextStreamToOpen = num + 1
 	s := m.streams[num]
@@ -167,5 +176,5 @@ func (m *incomingBidiStreamsMap) CloseWithError(err error) {
 		str.closeForShutdown(err)
 	}
 	m.mutex.Unlock()
-	m.cond.Broadcast()
+	close(m.newStreamChan)
 }